requirements.txt
anyio==4.6.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.16
django-extensions==3.2.3
djangorestframework==3.15.2
drf-yasg==1.21.7
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
inflection==0.5.1
packaging==24.1
pyTelegramBotAPI==4.23.0
python-telegram-bot==21.6
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
